# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Vocabulary Learning API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type WordDifficulty = typeof WordDifficulty[keyof typeof WordDifficulty];


export const WordDifficulty = {
  easy: 'easy',
  medium: 'medium',
  hard: 'hard',
} as const;

export interface Word {
  id: number;
  term: string;
  definition: string;
  /** @nullable */
  partOfSpeech?: string | null;
  /** @nullable */
  example?: string | null;
  /** @nullable */
  pronunciation?: string | null;
  /** @nullable */
  category?: string | null;
  difficulty?: WordDifficulty;
  /** 0-5 mastery scale */
  masteryLevel: number;
  reviewCount?: number;
  correctCount?: number;
  /** @nullable */
  lastReviewedAt?: string | null;
  /** @nullable */
  nextReviewAt?: string | null;
  createdAt: string;
}

export type WordInputDifficulty = typeof WordInputDifficulty[keyof typeof WordInputDifficulty];


export const WordInputDifficulty = {
  easy: 'easy',
  medium: 'medium',
  hard: 'hard',
} as const;

export interface WordInput {
  /** @minLength 1 */
  term: string;
  /** @minLength 1 */
  definition: string;
  partOfSpeech?: string;
  example?: string;
  pronunciation?: string;
  category?: string;
  difficulty?: WordInputDifficulty;
}

export type WordUpdateDifficulty = typeof WordUpdateDifficulty[keyof typeof WordUpdateDifficulty];


export const WordUpdateDifficulty = {
  easy: 'easy',
  medium: 'medium',
  hard: 'hard',
} as const;

export interface WordUpdate {
  term?: string;
  definition?: string;
  partOfSpeech?: string;
  example?: string;
  pronunciation?: string;
  category?: string;
  difficulty?: WordUpdateDifficulty;
}

export interface ReviewInput {
  correct: boolean;
  timeSpentMs?: number;
}

export type StatsByDifficulty = {
  easy?: number;
  medium?: number;
  hard?: number;
};

export interface Stats {
  totalWords: number;
  masteredWords: number;
  reviewedToday: number;
  currentStreak: number;
  totalReviews: number;
  accuracy: number;
  dueForReview?: number;
  byDifficulty?: StatsByDifficulty;
}

export interface Category {
  name: string;
  count: number;
  masteredCount?: number;
}

export type ListWordsParams = {
category?: string;
search?: string;
difficulty?: string;
mastered?: boolean;
};