# lib/api-zod/src/generated/types/reviewInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Vocabulary Learning API
 * OpenAPI spec version: 0.1.0
 */

export interface ReviewInput {
  correct: boolean;
  timeSpentMs?: number;
}