# lib/api-zod/src/generated/types/stats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Vocabulary Learning API
 * OpenAPI spec version: 0.1.0
 */
import type { StatsByDifficulty } from './statsByDifficulty';

export interface Stats {
  totalWords: number;
  masteredWords: number;
  reviewedToday: number;
  currentStreak: number;
  totalReviews: number;
  accuracy: number;
  dueForReview?: number;
  byDifficulty?: StatsByDifficulty;
}