# lib/api-zod/src/generated/types/word.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Vocabulary Learning API
 * OpenAPI spec version: 0.1.0
 */
import type { WordDifficulty } from './wordDifficulty';

export interface Word {
  id: number;
  term: string;
  definition: string;
  /** @nullable */
  partOfSpeech?: string | null;
  /** @nullable */
  example?: string | null;
  /** @nullable */
  pronunciation?: string | null;
  /** @nullable */
  category?: string | null;
  difficulty?: WordDifficulty;
  /** 0-5 mastery scale */
  masteryLevel: number;
  reviewCount?: number;
  correctCount?: number;
  /** @nullable */
  lastReviewedAt?: string | null;
  /** @nullable */
  nextReviewAt?: string | null;
  createdAt: string;
}