# lib/api-zod/src/generated/types/wordInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Vocabulary Learning API
 * OpenAPI spec version: 0.1.0
 */
import type { WordInputDifficulty } from './wordInputDifficulty';

export interface WordInput {
  /** @minLength 1 */
  term: string;
  /** @minLength 1 */
  definition: string;
  partOfSpeech?: string;
  example?: string;
  pronunciation?: string;
  category?: string;
  difficulty?: WordInputDifficulty;
}