# lib/api-zod/src/generated/types/wordUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Vocabulary Learning API
 * OpenAPI spec version: 0.1.0
 */
import type { WordUpdateDifficulty } from './wordUpdateDifficulty';

export interface WordUpdate {
  term?: string;
  definition?: string;
  partOfSpeech?: string;
  example?: string;
  pronunciation?: string;
  category?: string;
  difficulty?: WordUpdateDifficulty;
}